# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Bonus System API
  description: |
    API для управления бонусными программами с поддержкой мультитенантности и интеграцией с Telegram ботами.
    
    ## Основные возможности:
    - Управление проектами (тенантами)
    - Управление пользователями и бонусами
    - Webhook интеграция для внешних систем
    - Управление Telegram ботами
    - Аналитика и статистика
    - Реферальная программа
    - Многоуровневая система бонусов
  version: 1.0.0
  contact:
    email: support@saas-bonus.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.saas-bonus.com
    description: Production server
  - url: http://localhost:5006
    description: Development server
security:
  - ClerkAuth: []
...
paths:
...
  /api/projects/{id}/notifications:
    post:
      tags:
        - Telegram
      summary: Отправить расширенные Telegram-уведомления
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  maxLength: 4000
                imageUrl:
                  type: string
                  format: uri
                buttons:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      url:
                        type: string
                        format: uri
                      callback_data:
                        type: string
                parseMode:
                  type: string
                  enum: [Markdown, HTML]
                userIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Результат отправки уведомлений
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      sent:
                        type: integer
                      failed:
                        type: integer
        '400':
          description: Неверные данные
        '404':
          description: Проект не найден
        '429':
          description: Превышен лимит запросов
  /api/projects/{id}/users/{userId}/bonuses:
    post:
      tags:
        - Bonuses
      summary: Начислить бонусы пользователю
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AwardBonusRequest'
      responses:
        '200':
          description: Бонусы начислены
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  bonus:
                    $ref: '#/components/schemas/Bonus'
                  balance:
                    $ref: '#/components/schemas/UserBalance'

    get:
      tags:
        - Bonuses
      summary: Получить баланс пользователя
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Баланс пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  balance:
                    $ref: '#/components/schemas/UserBalance'
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT токен для аутентификации
  schemas:
...
    Bonus:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [PURCHASE, BIRTHDAY, MANUAL, REFERRAL, PROMO]
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
        isUsed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserBalance:
      type: object
      properties:
        currentBalance:
          type: number
        totalEarned:
          type: number
        totalSpent:
          type: number
        bonuses:
          type: array
          items:
            $ref: '#/components/schemas/Bonus'
...
    AwardBonusRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          minimum: 0
        type:
          type: string
          enum: [MANUAL, BIRTHDAY, PROMO]
          default: MANUAL
        description:
          type: string
        expiryDays:
          type: integer
...